+++ actix-files/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Opt-In filesize threshold for faster synchronus reads that allow for 20x better performance.
+- Add `{Files, NamedFile}::read_mode_threshold()` methods to allow faster synchronous reads of small files.
 - Minimum supported Rust version (MSRV) is now 1.75.
 
 ## 0.6.6